More big changes.
These are terrible commits, but right now the project feels like I have it
strewn all over the floor in pieces around me and I'm slowly putting things
into order. The parser is now mostly up and running, though I'm still shaking
out the test cases.

The next step will be to plug it in to the mask store.