Implement our own lr table generator.
I think that this is ultimately going to be easier than continuing to rely on
lark. The next component will be the parser for the lark files' grammar, but
luckily that can itself be implemented in the same parser. The ugly part will
be code generation. Maybe this is all a waste of time, but at least I'll have
learned a great deal.