Fix indexing issues in table generation.
Bamboozlingly this actually seems to be correct, at least visually comparing
the results to the expected ones from the Dragon Book (figure 4.42, page 266,
sec. 4.7.3 gives the table that this grammar should generate). I think that's
actually a working generator for LR parsing tables implemented? I can in the
near future copy out the expected result from the Dragon Book explicitly in
terms of the data structures I've used here, which are far from optimal I'm
sure. The next thing is to fix the interface with the parser module (and
optionally implement code generation so that we can actually read in and parse
the grammar files, whose grammar is itself defined as an LR grammar. Then once
that's up and going we can return to plugging that in to the lexer and so on.

I feel vindicated in my strategy of bypassing lark entirely and reimplementing
everything from scratch.