Experiment with new structure.
I'm quite tempted by doing the new version as a mono-lingual rust project and
forgetting all about the python bindings until the last possible moment, but
for now we're operating on the assumption that the python bindings are quite
important. It's my conviction that all the python will eventually wither away
as vestigial, except for whatever shims will eventually be needed to connect
between the completed rust version and the python interface we chose to
develop.

This will, before all is said and done, require plugging the rust part
into a language model directly, without going through python. Between candle
and tokenizers, it should be perfectly possible to do this. Tokenizers makes it
easy to access the vocabularies, which are the first thing we need; then the
models themselves can be gotten with candle-transformers, and we're in business.