Massage mask store into compliance.
We're getting dangerously close now: I've got to fix the tests for the mask
module. Then the only thing left for the rust component is the grammar
desugarer, which is going to be its own whole nightmare. It's not necessarily
easier, I don't think, to use Lark's existing implementation for this, but it
will serve as a strong point of reference for the implementation.